extension_server: Keep track of pending replies in the InstanceMap
As we can have at most one pending reply per instance, it makes sense to
keep the pending replies in the same struct as the instances map.

And with this change we avoid having to partially apply the IPC::Message in
the callback, as it is stored in the InstanceMap.