Remaove post_message callback from CreateInstance()
Instead of passing the same callback everytime an instance is created,
we may use the (already provided) SetPostMessageCallback() method.

This is preparation for the next commit, if we didn't this first we
would need to add another callback to CreateInstance to maintain consistency
which would increase the noise generated by the change.