[Extensions] XWalkExtensionModule create its instance
Now it's up to XWalkExtensionModule to create its instance in the
Client. This is a preparation for only creating an instance when the
extension is loaded.

As part of the change, the RemoteRunner abstraction was removed. In the
end, it wasn't bringing us much separation, especially now
ExtensionModule needs a way to create instances (which is Client job).

The runner had an interface that the ExtensionModule implemented, and
was used to tell it about messages from native. This interface was moved
to XWalkExtensionClient::InstanceHandler.