[Application][Tizen] Exit the application when the TERMINATE event is received
If we want the application to quit after it is terminated via the
Tizen task manager, we need to handle this event.

As Crosswalk will be following the unique name of the launcher process
and will kill the appropriate application when its launcher exits the
D-Bus session bus, simply calling exit() is enough.